User can now hide conversations in their profile
There is now a button below each conversation which makes a post request to the profile servlet.
Now this servlet handles two forms in its doPost function: the about me message change and hiding
conversations. Therefore, a switch statement is used to determine which functionality is needed.
Updated ProfileServletTests but still need to add a few more.

Current bugs:
*Can't un-hide conversations
*Anyone can hide your conversations.